Create a StratMan header file
